% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{computeSignificance}
\alias{computeSignificance}
\title{computeSignificance}
\usage{
computeSignificance(gg, log.p = TRUE, nquantiles = 5)
}
\arguments{
\item{gg}{A GenoGAM object.}

\item{log.p}{Should values be returned in log scale?}

\item{nquantiles}{The number of quantiles for base level computation.
Only effective if the GenoGAM object is of class GenoGAMList or
the underlying data is stored in HDF5 format. See details.}
}
\value{
An updated GenoGAM object, where the p-value slot is added.
}
\description{
The function computes positionwise p-values
}
\details{
In case of GenoGAMList the base level is computed as the median over
a list of chromosome data. This is achieved by computing a number of quantiles
for each chromosome and then compute the median on the complete set of quantiles
over all chromosomes. Similarly the same is done for the underlying HDF5 data,
but possibly with an even larger set of quantiles. That is because the data is read and
quantiles computed for each block to keep memory footprint low. Unless the organisms is
small, blocks have smaller size than chromosomes. Note, that this is an approximative
approach and the median is not necessarily closer to the overall 'real' median with
larger number of quantiles (unless in the extreme case, where each quantile is one data point).
5 quantiles (the default setting) is a number that gave good results in practice.

Note, that in case the data is stored in HDF5 format, the pvalue 'group' is added on hard drive.
That is, unlike any other function in R, where the input object is not changed, it actually
is in this case. If one wishes to have HDF5 data without the pvalue 'group', one has to
backup the HDF5 files prior to computation.
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
