% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMSetup-class.R
\docType{methods}
\name{dim,GenoGAMSetup-method}
\alias{dim,GenoGAMSetup-method}
\title{the dimension function}
\usage{
\S4method{dim}{GenoGAMSetup}(x)
}
\arguments{
\item{x}{A GenoGAMSetup object}
}
\value{
The four dimensions of the object (designMatrix rows, designMatrix
columns, experiment design rows, experiment design columns)
}
\description{
the dimension function
}
