% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{.absRanges}
\alias{.absRanges}
\title{A function to enhance absoluteRanges function to values beyond 2^31
by splitting the data and using own Coordinates class.}
\usage{
.absRanges(x)
}
\description{
A function to enhance absoluteRanges function to values beyond 2^31
by splitting the data and using own Coordinates class.
}
