% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_handler.R
\name{.createH5DF}
\alias{.createH5DF}
\title{initialize HDF5 dataset with specified dimensions and chunk}
\usage{
.createH5DF(seed, settings, d, chunk = NULL, what = c("fits", "coefs",
  "sumMatrix"))
}
\description{
initialize HDF5 dataset with specified dimensions and chunk
}
