% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMList-class.R
\name{.GenoGAMListFromHDF5}
\alias{.GenoGAMListFromHDF5}
\title{A function to create GenoGAMList from HDF5}
\usage{
.GenoGAMListFromHDF5(ggd, ...)
}
\description{
A function to create GenoGAMList from HDF5
}
