% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{.getChunkCoords}
\alias{.getChunkCoords}
\title{Function to establish chunk coordinates}
\usage{
.getChunkCoords(x)
}
\arguments{
\item{x}{An Coordinates object as the output of .getCoordinates}
}
\value{
The same object as x but with not overlapping ranges
which were cut at the center of the overhang
}
\description{
Function to establish chunk coordinates
}
