% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{.getCoordinates}
\alias{.getCoordinates}
\title{Function to retrieve the row coordinates as a list}
\usage{
.getCoordinates(x)
}
\arguments{
\item{x}{The GenoGAMDataSet object}
}
\value{
An integerList with the row numbers for each tile
}
\description{
Function to retrieve the row coordinates as a list
}
