% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{.getCoordinatesGGD}
\alias{.getCoordinatesGGD}
\title{Function to retrieve the row coordinates}
\usage{
.getCoordinatesGGD(x)
}
\arguments{
\item{x}{The GenoGAMDataSet object}
}
\value{
A Coordinates object specifying the row coordinates
of each tile
}
\description{
Function to retrieve the row coordinates
}
