% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{.getCoordinatesGGDL}
\alias{.getCoordinatesGGDL}
\title{Function to retrieve the row coordinates}
\usage{
.getCoordinatesGGDL(x)
}
\arguments{
\item{x}{The GenoGAMDataSetList object}
}
\value{
A Coordinates object specifying the row coordinates of
each tile
}
\description{
Function to retrieve the row coordinates
}
