% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{.getIdentifier}
\alias{.getIdentifier}
\title{get the identifier of the HDF5 files, that belong together to one dataset}
\usage{
.getIdentifier(path, fits = FALSE)
}
\description{
get the identifier of the HDF5 files, that belong together to one dataset
}
