% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{.getListedCoordinates}
\alias{.getListedCoordinates}
\title{Function to retrieve the row coordinates as a list}
\usage{
.getListedCoordinates(ind, rr)
}
\arguments{
\item{ind}{The GenoGAMDataSetList index}

\item{rr}{The GenoGAMDataSetList chromosome specific rowRanges}
}
\value{
An integerList with the row numbers for each tile
}
\description{
Function to retrieve the row coordinates as a list
}
