% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.linesearch}
\alias{.linesearch}
\title{the simple backtrack algorithm for linesearch
to ensure Wolfe conditions. See Nocedal, Algorithm 3.1}
\usage{
.linesearch(x, p, f, grad, alpha0 = 1, c = runif(1), rho = 0.9, ...)
}
\description{
the simple backtrack algorithm for linesearch
to ensure Wolfe conditions. See Nocedal, Algorithm 3.1
}
