% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_handler.R
\name{.makeFilename}
\alias{.makeFilename}
\title{make HDF5 name for the different files to be written}
\usage{
.makeFilename(dir, name, seed = NULL, split = " ", id = FALSE)
}
\description{
make HDF5 name for the different files to be written
}
