% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{.MetricsFunHDF5}
\alias{.MetricsFunHDF5}
\title{compute metrics for each tile}
\usage{
.MetricsFunHDF5(x, what, na.rm = FALSE)
}
\arguments{
\item{x}{The GenoGAMDataSet object}

\item{what}{A character naming the metric}

\item{na.rm}{Should NAs be ignored}
}
\value{
The metric value
}
\description{
compute metrics for each tile
}
