% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{.readRawData}
\alias{.readRawData}
\title{The intermediate function calling the correct
function to read in data based on the suffix and
computing the coverage from the GRanges output.
Returns a list of coverage Rles per chromosome}
\usage{
.readRawData(path, ...)
}
\description{
The intermediate function calling the correct
function to read in data based on the suffix and
computing the coverage from the GRanges output.
Returns a list of coverage Rles per chromosome
}
