% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{.setOptimalChunkSize}
\alias{.setOptimalChunkSize}
\title{set optimal chunk length if not provided
based on number of cores, samples, splines and
in case of HDF5, the block size}
\usage{
.setOptimalChunkSize(expDesign, design, ov, bpknots, hdf5)
}
\description{
set optimal chunk length if not provided
based on number of cores, samples, splines and
in case of HDF5, the block size
}
