% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{.subsetByOverlapsGGDL}
\alias{.subsetByOverlapsGGDL}
\title{underlying function to subset by overlaps}
\usage{
.subsetByOverlapsGGDL(query, subject, maxgap = -1L, minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"), invert = FALSE,
  ...)
}
\description{
underlying function to subset by overlaps
}
