% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMList-class.R
\name{.subsetByOverlapsGGL}
\alias{.subsetByOverlapsGGL}
\title{underlying function to subset by overlaps}
\usage{
.subsetByOverlapsGGL(query, subject, maxgap = -1L, minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"), invert = FALSE,
  ...)
}
\description{
underlying function to subset by overlaps
}
