% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{factorize}
\alias{factorize}
\title{Copied unchanged from Bill Venables package 'conf.design'
Version: 2.0.0
Published: 2013-02-23 15:18:29}
\usage{
factorize(x, divisors = primes(max(x)), ...)
}
\description{
Slightly changed the output
}
