% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGenoGAM-package.R
\name{fastGenoGAM}
\alias{fastGenoGAM}
\title{fastGenoGAM: A package providing a framework to analyse ChIP-Seq data}
\description{
fastGenoGAM: A package providing a framework to analyse ChIP-Seq data
}
