% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class_dev.R
\docType{class}
\name{GenoGAM-class}
\alias{GenoGAM-class}
\alias{GenoGAM}
\alias{design}
\alias{sizeFactors}
\alias{getSettings}
\alias{getFamily}
\alias{colData}
\alias{getParams}
\alias{getSmooths}
\alias{design,GenoGAM-method}
\alias{sizeFactors,GenoGAM-method}
\alias{getSettings,GenoGAM-method}
\alias{colData,GenoGAM-method}
\alias{getParams,GenoGAM-method}
\alias{getSmooths,GenoGAM-method}
\title{GenoGAM class}
\usage{
GenoGAM(...)

\S4method{design}{GenoGAM}(object)

\S4method{sizeFactors}{GenoGAM}(object)

\S4method{getSettings}{GenoGAM}(object)

\S4method{colData}{GenoGAM}(x)

\S4method{getParams}{GenoGAM}(object)

\S4method{getSmooths}{GenoGAM}(object)
}
\arguments{
\item{...}{Slots of the GenoGAM class. See the slot description.}

\item{object, x}{For use of S4 methods. The GenoGAM object.}
}
\value{
An object of the type GenoGAM.
}
\description{
This is the class the holds the complete model as well all hyperparameters
and settings that were used to fit it. It extends the RangedSummarizedExperiment
class by adding a couple of more slots to hold hyperparameters and settings.
The 'assays' slot holds the basepair fit and standard deviation. Additionally
all knot positions and beta coefficients will be stored in the 'smooths' slot
in order to be able to make use of the piecewise function that produces the
fit. For information on the slots inherited from SummarizedExperiment
check the respective class.

The GenoGAM constructor, not designed to be actually used, by the user.
Rather to be a point of reference and documentation for slots and how
to access them.
}
\section{Methods (by generic)}{
\itemize{
\item \code{design}: An accessor to the design slot

\item \code{sizeFactors}: An accessor to the sizeFactors slot

\item \code{getSettings}: An accessor to the settings slot

\item \code{colData}: An accessor to the factorialDesign slot.
It overwrites the inherited colData function, which points to the colData slot.

\item \code{getParams}: An accessor to the params slot

\item \code{getSmooths}: An accessor to the smooths slot
}}

\section{Slots}{

\describe{
\item{\code{family}}{The name of the distribution family used}

\item{\code{design}}{The formula of the model}

\item{\code{sizeFactors}}{The offset used in the model.}

\item{\code{factorialDesign}}{The factorial design used. The same as colData in the
GenoGAMDataSet}

\item{\code{params}}{All hyperparameters used to fit the data. The parameters
estimated by cross validation can also be found here. But the parameters
used in cross validation are in the settings slot.}

\item{\code{settings}}{A GenoGAMSettings object representing the global
settings that were used to compute the model.}

\item{\code{smooths}}{A data.table of knots and coefficients of the model}
}}

\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
