% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genogam.R
\name{genogam}
\alias{genogam}
\title{genogam}
\usage{
genogam(ggd, lambda = NULL, theta = NULL, family = "nb", H = 0,
  bpknots = 20, kfolds = 10, intervalSize = 20, regions = 20,
  regionSize = 4000, order = 2, m = 2)
}
\arguments{
\item{ggd}{The GenoGAMDataSet object to be fitted}

\item{lambda}{The penalization parameter. If NULL (default) estimated by
cross validation. So far only one parameter for all splines is supported.}

\item{theta}{The global overdispersion parameter. If NULL (default) estimated
by cross validation.}

\item{family}{The distribution to be used. So far only Negative-Binomial (nb)
is supported.}

\item{H}{The factor for additional first-order regularization. This should be
zero (default) in most cases. It can be useful for sparse data with many
zero-counts regions or very low coverage. In this cases it is advised to use
a small factor like 0.01, which would penalize those regions but not the ones
with higher coverage. See Wood S., Generalized Additive Models (2006) for more.}

\item{bpknots}{The spacing between knots. The lower the number, the more local
functions and the more sensitive the model with the trade-off of increased
computation time. In our experience going below 20 is not necessary even for
close double peaks.}

\item{kfolds}{The number of folds for cross validation}

\item{intervalSize}{The size of the hold-out intervals in cross validation.
If replicates are present, this can easily be increased to twice the fragment
size to capture more of the local correlation. If no replicates are present,
keep the number low to avoid heavy interpolation (default).}

\item{regions}{How many regions should be used in cross validation? The number
is an upper limit. It is usually corrected down, such that the total number of
models computed during cross validation does not exceed the total number of
models to compute for the entire genome. This is usually the case for small
organisms such as yeast.}

\item{order}{The order of the B-spline basis, which is order + 2, where 0
is the lowest order. Thus order = 2 is equivalent to cubic order (= 3).}

\item{m}{The order of penalization. Thus m = 2 penalizes the second differences.}
}
\value{
The fit as a GenoGAM object
}
\description{
The main modelling function.
}
\examples{
ggd <- makeTestGenoGAMDataSet(sim = TRUE)
res <- genogam(ggd, lambda = 266.8368, theta = 2.415738)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
