% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMSettings-class.R
\docType{class}
\name{GenoGAMSettings-class}
\alias{GenoGAMSettings-class}
\alias{GenoGAMSettings}
\title{GenoGAMSettings}
\usage{
GenoGAMSettings(...)
}
\arguments{
\item{...}{Any parameters corresponding to the slots and their possible
values.}
}
\description{
This class is designed to store global settings for the computation of the
GenoGAM package

The Constructor function for GenoGAMSettings is merely a wrapper for
new("GenoGAMSettings", ...)
}
\details{
Center can have three values: TRUE, FALSE, NULL. TRUE will
trigger the center function, FALSE will trigger the use of the entire
fragment. NULL should be used in case a custom process function is used.
In case a custom function is used, it has to satisfy the following:
It has to handle a GAlignments object as input and output a 
GRanges object of regions, e.g. fragments. This regions are in turn used
to compute the coverage via the IRanges::coverage function. Note, that there
is a difference between the GAlignments object in the single and paired end
case.
}
\section{Slots}{

\describe{
\item{\code{center}}{A logical or NULL value to specify if the raw data should be
centered, i.e. only the midpoint of the fragment will be used to
represent its coverage. See details.}

\item{\code{chromosomeList}}{A character vector of chromosomes to be used.
NULL for all chromosomes.}

\item{\code{bamParams}}{An object of class ScanBamParam.
See ?Rsamtools::ScanBamParam for possible settings. Usually used to set
specific ranges, to read in.}

\item{\code{processFunction}}{A custom function on how to process raw data. Not
used if center is TRUE/FALSE. This is not intended for the user, but if
needed anyway, see details.}

\item{\code{optimMethod}}{The optiomisation method to be used in cross validation.
See ?optim for a complete list.}

\item{\code{optimControl}}{Settings for the optim function. See ?optim for a 
complete list.}
}}

\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
