% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class_dev.R
\docType{methods}
\name{getFamily,GenoGAM-method}
\alias{getFamily,GenoGAM-method}
\title{describeIn GenoGAM An accessor to the family slot}
\usage{
\S4method{getFamily}{GenoGAM}(object)
}
\description{
describeIn GenoGAM An accessor to the family slot
}
