% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMSetup-class.R
\docType{methods}
\name{length,GenoGAMSetup-method}
\alias{length,GenoGAMSetup-method}
\title{the length function}
\usage{
\S4method{length}{GenoGAMSetup}(x)
}
\arguments{
\item{x}{A GenoGAMSetup object}
}
\value{
The length of the object as the product of all dimensions
}
\description{
the length function
}
