% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{makeTestGenoGAMDataSetList}
\alias{makeTestGenoGAMDataSetList}
\title{Make an example /code{GenoGAMDataSet}}
\usage{
makeTestGenoGAMDataSetList()
}
\arguments{
\item{sim}{Use simulated data (TRUE) or test data from a real experiment}
}
\value{
A /code{GenoGAMDataSet} object
}
\description{
Make an example /code{GenoGAMDataSet}
}
