% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{primes}
\alias{primes}
\title{Copied unchanged from Bill Venables package 'conf.design'
Version: 2.0.0
Published: 2013-02-23 15:18:29}
\usage{
primes(n)
}
\description{
Copied unchanged from Bill Venables package 'conf.design'
Version: 2.0.0
Published: 2013-02-23 15:18:29
}
