% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGenoGAM-package.R
\docType{package}
\name{fastGenoGAM}
\alias{fastGenoGAM}
\alias{fastGenoGAM-package}
\alias{fastGenoGAM-package}
\title{fastGenoGAM: A package providing a framework to analyse ChIP-Seq data}
\description{
fastGenoGAM: A package providing a framework to analyse ChIP-Seq data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gstricker/GenoGAM}
  \item Report bugs at \url{https://github.com/gstricker/GenoGAM/issues}
}

}
\author{
\strong{Maintainer}: Georg Stricker \email{georg.stricker@in.tum.de}

Authors:
\itemize{
  \item Alexander Engelhardt \email{alexander.engelhardt@ibe.med.uni-muenchen.de}
  \item Julien Gagneur \email{gagneur@in.tum.de}
}

}
